Templated OctTree Class; Boundary Intersect Function
OctTree class uses a generic type for its contained 'objects'.
Started writing a function to get 'objects' within the boundary of an OctTree.
Added function to figure out if two boundaries are intersecting.
Made GetMatrixRow function public.

+++ Application/include/Application.h
@@ -26,11 +26,11 @@ public:
 
 	inline void SetBoidCount(unsigned int a_boidCount);
 
-	inline int GetBoidCount() const;
+	inline unsigned int GetBoidCount() const;
 	inline unsigned int GetMaximumBoidCount() const;
 
 private:
-	int m_uiBoidCount;
+	unsigned int m_uiBoidCount;
 	const unsigned int mc_uiMaximumBoidCount;
 	float m_fMarkerZOffset;
 	const float mc_fMimimumMarkerZOffset;
@@ -45,10 +45,10 @@ private:
 
 void Application::SetBoidCount(unsigned int a_boidCount)
 {
-	m_uiBoidCount = (int)a_boidCount;
+	m_uiBoidCount = a_boidCount;
 }
 
-int Application::GetBoidCount() const
+unsigned int Application::GetBoidCount() const
 {
 	return m_uiBoidCount;
 }
